pool: Fix timeout behavior of HSM requests
Timeouts in the pool were implemented from the request creation time rather
than the request activation time. This is a regression compared to 2.6 and
leads to problems when requests are queued: Since reqeusts are executed in FIFO
order, the requests that are activated are always the oldest requests.  Since
those older than the deadline have been removed, the oldest requests left are
often close to the deadline and will be cancelled shortly after activation.
This leads to a lot of churn with no real progress.

The intended implementation of timouts for HSM requests was never finished: The
intention was the whoever requests an HSM operation (in particular stage) would
specify a lifetime. E.g. SRM allows this. This would be the deadline and an
advanced provider could use this to schedule requests. Specific providers could
implement their own timeouts for particular operations, e.g.  callouts to
scripts. This was the reason for why the deadline was measured from the
creation time. Currently however only the latter form of timeout is implemented
and not within the provider but in the framework. Eventually this needs to be
cleaned up. For the time being this patch changes the deadline to be calculated
from the activation time.  Requests in the queue never time out.

This implementation is to be considered temporary (although without a limit on
when it will be changed).

Target: trunk
Require-notes: yes
Request: 2.12
Request: 2.11
Request: 2.10
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8620
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8224/