srm: Further optimize SRM scheduler
In case there are less requests in the queue than we are still allowed to
start, the need to apply any prioritisation is less. Since the prioritisation
is very expensive, this patch skips it entirely until the queue becomes longer
than the number of empty task slots. This is mostly a workaround for the slow
prioritization. A future patch for master should reimplement how requests
are prioritised.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.12
Request: 2.11
Request: 2.10
Issue: #1533
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: https://rb.dcache.org/r/8241/