srm: Fix erroneous message queue overflow issues
The SRM uses a ThreadPoolExecutor with up to 100 threads and no queueing (queue
capacity is 1). The intention was that the bounded queue should cause the SRM
to dynamically scale it's number of threads and allow 100 messages to be
submitted "in bulk" before messages are dropped. This doesn't work though as
the messages are dropped as soon as a single unconsumed message is in the queue
- even if there are plenty of idle threads.

This patch extends BoundedExecutor to accept a maximum queue size and updates
the SRM to use this thread pool instead. BoundedExecutor has a more intuitive
model in which it first scales the number of threads and then begins queueing
(ThreadPoolExecutor is the other way around).

The patch obsoletes srm.cells.limits.threads.min,
srm.cells.limits.threads.max-idle-time, and
srm.cells.limits.threads.max-idle-time.unit.

Target: trunk
Require-notes: yes
Require-book: no
Issue: #1535
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: https://rb.dcache.org/r/8242/