libs: Update third party libraries
Mostly minor updates, except for the Berkeley DB update. That library is used
by pools and the new version introduces a new on disk format.  The update is
forward compatible but not backwards compatible. That means that sites can
upgrade without problems, but to downgrade the meta data has to be converted to
the file format before downgrade and then back to the berkeley db format after
downgrade. Changelog for Berkeley DB can be found here:

    http://docs.oracle.com/cd/E17277_02/html/changelog.html

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>