common: Extend execution time gauge proxy to asynchronous results
The monitoring proxy can wrap classes with counters and execution time gauges.
It tracks number of calls, failures and execution times, but is fairly useless
if the implementation is asynchronous and returns a Future.

This patch extends the proxy to recognize ListenableFuture return values. The
proxy adds its own listener to intercept the asynchronous result.

This makes the 'print srm counters' output of the SRM much more useful.

Target: trunk
Require-notes: yes
Require-book: no