srm: Add cache for srmGetSpaceTokens
Motivation:

An FTS job refers to space reservations by description rather than tokens.
Therefore FTS will query the available space tokens ahead of every transfer.
Sice this is not a schedulable operation, a Jetty thread is occupied for the
duration of the lookup.  The list of tokens rarely changes.

Modification:

Introduces a cache in the SRM for the space token lookup. The cache has a
relatively short life-time and implements asynchronous refresh.

Also changes the respose to srmGetSpaceTokens in case the space manager is
disabled: Previously we would report that the feature was unsupported. Now we
provide an empty list of tokens (which is the correct answer if space
maangerment is not supported).

Result:

Faster srmGetSpaceTokens processing.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8276/