spacemanager: Allow unowned files in reservations
Motivation:

Space manager tracks the "owner" of a file in a reservation. This owner is
unused, yet the code insist on there being an owner. Before 2.13, an owner can
only be an FQAN or user name and thus uploads to reservations fail with
protocols not providing either. NFS is such a protocol.

Most of the code doesn't care that there is no owner, but a specific utility
class, VOInfo, rejects a null value.

Modification:

VOInfo is updated to allow null values for group. The implementation of
hashCode and equals is updated to work with null values. This also fixes a
potential issue with null values for role which were allowed but would have
failed.

Result:

Fixes #1622.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Issue: #1622
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8294/