pnfsmanager: Inherit ACLs on upload with SRM
Motivation:

The SRM uses temporary upload directories, but pnfsmanager fails to copy the
ACLs of the actual target directory to the upload directory.  Thus the newly
uploaded file will inherit the ACLs from the base upload directory rather than
the taget directory.

Modification:

The name space provider copies the ACLs of the target directory to the
temporary upload directory. Chimera had to be modified to accept ACLs on mkdir
to efficiently set those. The setACL code path in Chimera got change to avoid
an unecessary reset of the ctime when the ACLs are not changed or set on a
newly created directory.

Result:

Fixes #1639

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8303/