webdav: Fix proxied upload with chunked encoding
Motivation:

A WebDAV door may proxy uploads. When relaying the data to the pool, the door
may send the expected file size to the pool in an HTTP header or use chunked
encoding. Which it chooses depends on whether the file size is known. The file
size may be know because the client itself submitted the file size in an HTTP
header, or because the client negotiated the transfer with SRM.

Since Chimera defaults to a file size of zero, the door will falsely set that
as the expected size even when no size is known and the client is using chunked
encoding. Such an upload will fails. ARC is a client using chunked encoding
for upload.

Modification:

This bug was previously fixed in 2.12. The fix was too intrusive to be
backported.  This patch makes a local fix in the WebDAV door to detect the
cases in which a client provided a size through either SRM or through an HTTP
header.

Result:

Fixes proxied chunked upload through HTTP. Fixes compatibility with ARC.

Target: 2.11
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8311/