infoprovider: remove single SRM instance limitation
Currently, if dCache has multiple SRM instances then they are
published, but with certain fields containing the same information.
Most criticially, the URL of the endpoint is hard-coded to be the
official ID of the instance.

With this patch, the actual FQDN of the endpoint is published.  For
most sites this will bring no change: the FQDN of the SRM endpoint and
the official ID are the same.

However, it will allow sites to publish multiple SRM endpoints.

Target: master
Patch: https://rb.dcache.org/r/8321/
Acked-by: Tigran Mkrtchyan
Acked-by: Gerd Behrmann
Require-nodes: yes
Require-book: yes
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10

Conflicts:
	skel/share/info-provider/glue-1.3-defn.xml

Conflicts:
	skel/share/info-provider/glue-1.3-defn.xml
	skel/share/info-provider/glue-2.0-defn.xml