webdav: provide a human-friendly size as file-size hint
The webdav service's HTML directory listing provides the file size
in bytes.  While accurate, this is not always the easiest way for
humans to appreciate the size of a file.

This patch makes the file size available to the template language in
different formats.  Currently two formats are provided: the number of
bytes as a simple integer and a "human friendly" string describing
the size with SI binary prefix and "B" (e.g., "MiB").

The default template is updated to show the human-friendly string
as a hint, typically shown when the mouse hovers over the table
element.

Target: master
Patch: https://rb.dcache.org/r/8337/
Acked-by: Gerd Behrmann
Requires-notes: yes
Requires-book: yes