srm: Refactor scheduler to avoid unused queues and per user counters
Motivation:

A recent patch introduced a new SchedulingStrategy abstraction. After that
patch it is no longer necessary to keep detailed per user statistics about
the number of jobs, nor is it necessary to maintain separate queues for
the TQUEUD and RQUEUED states.

Modification:

Introduces a single multimap to track jobs in each of the non-final states.
This allows us to print the jobs in TQUEUED and RQUEUED and it provides
counts for the number of jobs in each state.

Result:

A fair amount of legacy code is removed. The use of a multimap avoids
the O(n) remove operation of ModifiableQueue.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.13
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8339/
Bug: #1533
(cherry picked from commit edfa418628bfb53fc7e90b620abaab17568015d1)