(2.12) gplazma-xacml: add GlobusPrincipal (DN) to identified principals
A refactoring of the GsiFtpV1Door code from 2.11 to 2.12 changed
the presence of the DN ins the session log from optional to required.

As such, the XACML plugin, which has only ever added Login, UserName
and GID principals to that set, currently fails (see error below).

This patch adds the requisite principal.  A slightly different version
of the patch will be required for 2.12, as the VOMS validator class
has changed since then.

Target: 2.12
Patch: 8355
Require-notes: yes
Require-book: no
Acked-by:  Paul Millar