srm: Add transfer strategy plugin mechanism
Motivation:

dCache provides a pluggable mechanism for controlling SRM requests in the
in progress state, but it doesn't have any support for scheduling the
transition into the ready state (i.e. when the turl is handed out to the
client). This is however often the place where sites see contention as
the number of TURLs it limited.

Modification:

Adds a pluggable TransferStrategy mechanism. Two implementations are provided:

- fcscs, first come first server matches the old behaviour and hands out TURLs
  until a limit is reached.

- fair-share, implements a fair-share logic similar to inprogress-fair-share
  scheduling algorithm. Assuming a party has requests queued, slots will be
  held available to guarantee that clients fair share.

Note that the transfer strategy is not a scheduler as it can only provide
a yes/no answer when the client requests the TURL.

Since fair-share is the default scheduling algorithm, fair-share is the new
default for the transfer-strategy too.

Result:

Lower risk that one VO steps on the toes of another. Fixes #1691.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8360/