xrootd: Add session to access log
Motivation:

Netty processes all requests on a thread pool associated with the channel
context. When writing to a channel from another thread, the request is
first moved to the appropriate thread pool. In doing so the CDC is lost,
preventing the session to be logged to the access log.

Modification:

Wraps the response write in a runnable submitted on the appropriate
thread pool. Netty detects the call is from the correct thread and will
not resubmit the write itself. This gives us a chance to inject the
correct CDC.

Result:

The xrootd door access log logs the correct session. Fixes #1712.

Target: trunk
Request: 2.13
Request: 2.12
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8370/