gplazma: Fix VOMS plugin regressions
Motivation:

We recently upgraded to the VOMS Java API version 3. In the process code
for seting the voms directory and CA directory was lost, as well as
caching of the voms and CA directory content. As a result, current
dCache 2.13 is almost useless, as the performance overhead of reloading
the certificates for every login cripples throughput.

Modification:

Restore pre-instantiation of the components that load the voms
and ca directories.

It looks like VOMSACValidator itself should be thread safe too, but
currently I keep the old logic of reinstanting it for every request.
It appears the overhead of doing this is minimal.

Result:

Let the voms plugin respect the vomsdir and capath settings, and
remove a serious bottleneck from voms authentication.

Target: trunk
Requet: 2.13
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8380/