chimera: Fix path resolution on delete
Motivation:

When deleting a path /a/b/c ending in a symlink (i.e. c is a symlink),
the name space provider follows the symlink when verifying permissions.
This may mean that the request is denied, e.g. if following the symlink
is not allowed for the subject or if the object being linked to has an
ACL preventing deletion. This is wrong as only the symlink is deleted.

The actual delete happens on the symlink.

The current code is also suboptimal in that the path is actually resolved
three times.

Modification:

Resolves the path to the parent once (following symlinks) and then
does the permission checks on the element in that directory.

Result:

Correct permission checks and reduced overhead for file deletion.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8382/
(cherry picked from commit b37e0eb059f22bd824d24043acf4a9698115480d)
(cherry picked from commit 97a5085eb79f9290c3af92e00d8925879ee33555)