pnfsmanager: Minor refactoring of access mask check for delete
Motivation:

PnfsManager requests can provide an access mask indidicating additional
permissions to check as part of the operation.

Modification:

Rearranges the access mask check for file deletion to make use of a PNFS ID
in the request whenever it is present.

Result:

Avoids an extra path lookup in case both an PNFS ID and a path is present
and an access mask was specified.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8387/