chimera: Let SRM respect setgid on upload
Motivation:

SRM instructs the pnfs manager to create temporary upload directories for each upload.
The pnfs manager fails to check the setgid bit on the target directory and thus the
group is not inherited when setting the setgid bit.

Modification:

Check if the setgid bit is set and if so copy both that bit and the group of
the target directory.

Also remove the authorization check on commit and cancel: Indepedently of the setgid
problem, this is necessary as the owner and group of the target directory may be
different from the uid and primary gid and thus the check is erroneous. The setgid
behaviour just happens to run into this flaw as it is one way in which the group may
become different than the primary gid.

Result:

Fixes #1747.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8388/