admin: Fix cell name tab completion
Motivation:

Admin shell tab completes cell names. When the name to be completed does not
contain an @ symbol, the admin shell completes from the list of well known
names and all local cells of the connected domain. This completion however has
a bug that is triggered when there is more than one hob from the admin service
to the connected domain (eg admin is not in dCacheDomain and the shell is
connected to another domain that is not the dCacheDomain either). In this
case the local cells from dCacheDomain are used for completion.

Modifcation:

The bug is caused by using the current address of the cell path to the
connected cell. That happens to be the next hop, which is commonly
dCacheDomain. The fix is to use the last address of the path (the destination
address).

Result:

Correct tab completion for multi-domain deployments.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.13
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8389/