pnfsmanager: Remove uid,gid and mode from upload path creation
Motivation:

The PnfsCreateUploadPath message to create temporary upload directories carries
uid, gid, and mode fields to allow setting custom values. This is copied from
the PnfsCreateEntryMessage, but is otherwise not utilized.

There are several problems with the implementation that would manifest themselves
if this functionality would be used:

- The automatic creation of parent directories would get the specified mode field
  too and thus likely lack the x-bit. Thus those directories would be unwritable.

- The temporary upload directory would get the specified uid, gid and mode and
  would likely lack an x bit too - or would not be writable by the user as the
  owner is different from the user uploading the file.

The only reasonable use case I can think of is if SRM wants to set the stage
for an anoymous upload (e.g. plain HTTP) and thus wants to create a publicaly
writable upload directory. However the current implementation was not useful
for that either as the mode was applied to the target file too. If this
functionality is needed in the future, we have to implement it at that point.

Modification:

Since the functionality was unused, this patch removes those fields. UID, GID and
mode is now derived from the subject and the parent directory.

Result:

No observable change, except code that doesn't contain non-functional unused bits.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8392/