ftp: consolidate transaction-log creation
The FTP door supports a transaction log.  The initial log entry is delegated to the
concrete FTP instances, despite this initial entry being broadly the same.  This leads
to non-DRY code.

This patch consolidates this by moving the creation and initial line into
AbstractFtpDoorV1.  The name of the door (as logged in the transaction log) becomes
part of the constructor and only determining the user's name is overridden, and only
by the GSI-door.

There should be no user- or admin- visible changes from this patch.

Target: master
Patch: https://rb.dcache.org/r/8411
Acked-by: Tigran Mkrtchyan
Requires-notes: no
Requires-book: no