srm: Add authorization to put done and abort requests
Motivation:

For uploads, pnfsmanager used to authorize the completion or cancellation calls
based on the UID and the GID. A fix to a bug related to the setgid bit forced
those authorization checks to be removed (see https://rb.dcache.org/r/8388/).
This left those operations unauthorized.

Modification:

Adds authorization checks in the SRM for put done, abort files and abort calls.
Only clients with the same UID as the owner of the original request and clients
which have the primary GID of the owner of the original request as one of their
GIDs can make those calls.

Result:

The SRM returns SRM_AUTHORIZATION_FAILURE on srmPutDone, srmAbortFiles, and
srmAbortRequests if the client is not authorized to make these calls.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8436/