chimera: fix nameof and pathof for paths containing unicode
chimera pathof and nameof commands do not handle paths with unicode characters correctly, because it treats the paths as regular strings while the length of the string and the number of bytes occupied by the string differ.
This patch fixes this.

Ticket:
Acked-by: Paul
Target: master
Target: 2.13
Target: 2.12
Target: 2.11
Target: 2.10
Require-book: no
Require-notes: yes

Conflicts:
	modules/chimera/src/test/java/org/dcache/chimera/BasicTest.java