pnfsmanager: Use a single shared request queue for threads
Motivation:

Pnfs manager has multiple processing threads. Chimera and even more so PNFS
before it had concurrency issues. Due to this we deterministically mapped
requests for a particular file system object to queues such that all requests
on the same object would be processed sequentially.

With the recent changes to transaction management for Chimera, the concurrency
problems have disappeared and there is no longer a need to maintain several
requests queues.

Modification:

We now use a single request queue per thread group. The thread group concept is
kept as it allows the load to be split over multiple backend databases, using
separate queues and separate threads for each (although this is currently not
supported by Chimera).

The folding logic is dropped for several reasons. With multiple threads
processing requests on the same queue, it becomes difficult for those threads
to scan the queue without conflicting with each other. Moreover, there is
nothing preventing a concurrent read from dequeueing a message that could be
folded with the request the current thread just processed. Finally, the linear
scan - and the O(n) per element remove of LinkedBlockingQueue (yes, even remove
on the Iterator of LinkedBlockingQueue is O(n)) - would have become even more
expensive with a deeper queue. It is my hope that the introduction of a TTL for
pnfs manager messages eliviates some of the need for request folding. The low
level support in the request messages themselves has not been removed and can
be reused if we find a way to do so efficiently in pnfs manager.

Result:

Using a single shared queue allows the load to be balanced among all request
threads.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8452/