chimera: Fix two race conditions in file deletion
Motivation:

The file deletion code contained two race conditions:

- Upon deletion, the link count of the file to be deleted is read.
  Subsequently, if it was 1 the inode is deleted. If the file had two hard
  links that both get deleted at the same time, both threads would read a link
  count of 2 and none of them would delete the file, although both would unlink
  it. This would result in an inode without any links.

- The call to remove the link to the file is by parent inode and file name. The
  subsequent decrease and deletion of the file is by inode. There is nothing
  preventing a concurrent transaction from renaming the file and creating another
  one with the same name. The delete would then delete the newly created file (as
  the delete happens per name rather than inode), yet it is the link count of the
  original (and now renamed) file that gets decreased. If this was the only link
  to the file, attempting to delete the inode will trigger a foreign key violation
  and the transaction gets rolled back. However if there are more than one hard
  link to the file, the code will not actually remove the inode and thus no
  violation occurs. The result is that we are left with an orphaned inode and
  another inode that has a wrong link count.

Modification:

The primary change is to delete by both name and inode. Secondly, we rely on the
fact that an inode is only deleted if the link count is zero and thus that
we do not need to read the link count ahead of time.

A number of sanity checks are added to the relevant code.

Unit tests are added; these do not check the race condition (not easily possible
in unit tests), but verify the remainig behaviour.

Result:

Both race conditions are fixed.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8455/