chimera: Fix races in move/rename logic
Motivation:

The move logic had several race conditions caused by looking up the inode
in a READ_COMMITTED transaction and subsequently performing changes based
on this inode without verifying whether something related to this inode
had changed in a concurrent transaction:

- The code checks whether the destination name exists already; if it is the
  same file as the source the entire move is skipped. Otherwise the existing
  destination is deleted; the comparison is by inode, yet the delete is by
  name and thus there is no guarantee that a concurrent transaction could
  not have deleted the file already and added a link instead.

- Similarly, the source could be deleted concurrently and a link to the
  target could be established. Since the subsequent move is by name rather
  than inode, this too would have lead to an erroneous result.

- The source and destination inode types are compared and overwrite is
  only allowed for the same file system object type (i.e. a file may not
  overwrite a directory). This too is suspect to races as the delete happens
  by name and thus a concurrent transaction could have deleted the destination
  file and replaced it with a directory; in such a case the move would
  overwrite a directory.

Modification:

The solution is to constrain the delete and the remove to the particular
inode we already resolved the path to. If entry changed in the meantime,
we repeat the move.

To implement this, both the move and the remove logic had to be refactored.

Several consistency checks have been added to provide fail fast behaviour
in case an update doesn't update the expected number of rows.

Result:

Race conditions fixed.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8456/