pnfsmanager: Fix races in file deletion
Motivation:

The ChimeraNameSpaceProvider needs to verify permissions before calling into
Chimera to delete a file. Yet the delete call into Chimera is by name and thus
doesn't protect against concurrent replacement of the file.  Thus the file
being deleted could be different from the one which we authorized.

Modification:

Change the Chimera interface to accept an inode in addition to the directory
and file name. The entry is only deleted if the entry points to the given
inode.

Result:

Fixes the bug and also eliminates some redundant lookups of the inode.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8458/