chimera: Drop trigger to copy tags
Motivation:

For postgresql a trigger on inserting the '..' entry is used to copy the tags
from the parent directory to a newly created directory. This trigger saves
one round-trip from the client to the server upon creating directories, but
does not otherwise reduce the load on the database.

Not all directories should inherit the tags upon creation: Temporary upload
directories should not and the postgresql driver for Chimera has code to
delete any tags copied by the trigger. This adds an extra round-trip to the
server and also adds extra load on the database.

On a system with SRM, upload directory creation must be assumed to be by far
the most common form of directory creation. Even for systems without SRM,
the trigger did not reduce the load on the database - it merely reduced the
latency. If latency for directory creation is critical, using a stored
procedure for the entire mkdir operation would be by far better.

Modification:

Dropped the trigger and restored the base driver's functionality to copy
the directory tags.

Result:

Reduced database load on upload directory creation. If regular directory
creation latency is critical, a stored procedure for mkdir should be added in a
future patch and called explicitly from the postgresql chimera driver.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8462/