chimera: Fix hard link creation in the pressence of ACLs
Motivation:

ACLs are inherited from the parent directory upon inode creation. Our previous
implementation implemented this using database triggers watching for insertion
into the t_dirs table. This would cause hard link creation to fail if the
parent directory of the hard link contained inheritable ACL while the linked to
inode already had an ACL. This is because the hard link would insert an entry
into t_dirs and the trigger would erroneously try to copy the ACL.

Modification:

Removes the ACL database trigger and replaces it with explicit queries to copy
the ACLs upon file and directory creation. Thus we avoid the erroneous copy
of the ACL when a hardlink is created.

The query to copy the ACLs is implemented for each supported database backend
as there is no standard syntax for bitwise operations in SQL. There is a fallback
implementation in Java for other databases.

Result:

Hard link creation works in the pressence of ACLs. ACL inheritance works for
all supported databases.

This change adds an extra round-trip to the server, but at the same time
reduces the load on the database as there is no longer a need to lookup the
inode type whenever an entry is added to t_dirs. For directories, three such
entries are created, thus saving three lookups. If the extra roundtrip to
the server is critical, a stored procedure for create entry operation should
be created an called explicitly.

For creating directories with a prespecified ACL (eg temporary upload
directories), the patch further saves a redudant copy + delete cycle for the
ACL of the parent directory.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8463/