chimera: Do not store '.' and '..' entries in the database
Motivation:

At NDGF, approx. 29e6 out of 88e6 entries in t_dirs are for '.' and '..'. These
entries are entirely redundant, as the same information can be extracted from
other data in the database.

For the same reason, many real file systems do not store these entries on disk
either, opting instead to simulate them in software.

Modification:

Drop the '.' and '..' entries from the database and update Chimera not to
create or expect these entries. Listing and lookup code is updated to detect
these special names and do the equiavalent lookup.

In a future patch we may consider not even adding these entries to the list
output as it appears that both the NFS 4 interface and the pnfsmanager
interface filter those entries from the result anyway.

The link count interpretation for directories is unchanged. Chimera directories
use the link count to track the number of entries in the directory and as
before the virtual '.' and '..' entries are included in the link count.

Result:

This patch would lead to a 33% reduction of the t_dirs table at NDGF. Although
disk space isn't critical, being able to cache more relevant data in memory is
important. Also the reduced index sizes for t_dirs should have a positive
impact on lookup performance.

Upon upgrade, the '.' and '..' entries will be automatically removed. Schema
rollback is supported to add those entries back.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8467/