ftp: prevent execution of most commands when unwrapped
This patch adds an additional check to verify that (almost all) commands are
wrapped when executed by a GSS-backed door.  The commands for establishing
a context and for wrapping a command are exempted from this.

The CCC command is still supported; this downgrades the control channel to
allow unencrypted communication.

Target: master
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Request: 2.6