srm: Cache the vomsdir and ca stores
Motivation:

The SRM uses a generic utility class in gPlazma2 to extract FQANs from a
certificate chain. This utility class repeatedly reloads the files in vomsdir
and the ca certificates. In particular the latter is associated with
significant overhead.

This is a regression in 2.13.

Modification:

Embeds the FQAN extraction code directly into the relevant places in the SRM,
which allows us to cache the store material. The utility class is removed
entirely as the only users were the SRM and the DCAP doors.

Result:

Faster SRM and fewer dependencies between our modules.

Target: trunk
Require-notes: no
Require-book: no
Request: 2.13
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8474/