srm: Fix credential delegation
Motivation:

We use a subclass of SSLEngine to implement the GSI delegation. This subclass
relies on the client sending the proxy certificate in a single SSL frame. This
is not always the case: The classic example is when the proxy is very large and
doesn't fit in a single frame, but other cases of the SSL engine splitting the
payload into multiple frames have been observed.

Modification:

Detects incomplete certificates by parsing the ASN1 object and detecting an
EOFException. In that case another SSL frame is read and the operation is
repeated.

Although not the most elegant solution, the alternative would have been to
implement (or copy) a full BER parser to determine the length of the encoded
certificate.

Also fixes a bug in GsiEngine in which after completed delegation it reads
another SSL frame from the input. As a result it never returns a handshake
status of FINISHED. The SSLEngine specification states that unwrap should
only consume a single SSL frame per invocation.

Result:

Improved compatibility.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8480/
(cherry picked from commit 604a0babd4c780382eee8938be58dba2300bc564)