Relocate GSI related classes to javatunnel
Motivation:

If we are to make use of our own GSI implementation in the GridFTP and SRM client,
we cannot have those classes in the main dCache package.

Modification:

Relocates the GSI classes to the javatunnel package. That package thus becomes
a generic home for encryption related stuff and probably should be renamed in
a subsequent patch.

It introduces some dependencies from javatunnel to a couple of Jetty packages
(util and io). We use these in our GSI implementation and they are not as
such specific to Jetty. We may consider refactoring the code at some point in
the future to get rid of these dependencies.

Result:

Enable the use of our GSI implementation in clients without pulling in all
of dCache as a dependency.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8483/